pyproject.toml
[project]
name = "gsay"
version = "0.0.1"
description = "ggwave server"
# readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "fastapi==0.101.0",
    "fasteners==0.18",
    "ggwave==0.4.2",
    "numpy==2.2.6",
    "paho-mqtt==2.1.0",
    "pydantic==1.10.19",
    "python-dotenv==1.0.1",
    "soundcard>=0.4.4",
    "soundfile==0.13.1",
    "uvicorn==0.23.2",
]

[project.scripts]
gsay = "gsay:gsay.main"
gserver = "gsay:gserver.main"
glisten = "gsay:glisten.main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
